chore: Moved utility class to engine folder
