+++ ts/engine/page-management.ts
@@ -1,5 +1,5 @@
 import Page from "./page";
-import {Utility, ScrollDirection} from "../utility.js";
+import {Utility, ScrollDirection} from "./utility.js";
 
 export class PageManagement {
     // A class that manages and controls all the pages
