+++ ts/engine/utility.js
