+++ ts/engine/utility.ts
